Relocate config classes into own package
